move NavBar.jsx to index.jsx
